Move N50 calculation to utils. Update documentation
